requirements.txt
APScheduler==3.6.0
asn1crypto==0.24.0
beautifulsoup4==4.7.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==1.1.15
cryptography==2.4.2
Flask==1.0.2
future==0.17.1
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.1.0
pika==1.0.1
pipdeptree==0.13.1
pycparser==2.19
pymongo==3.7.2
python-dateutil==2.6.0
python-telegram-bot==11.1.0
pytz==2018.7
PyYAML==3.13
requests==2.21.0
six==1.12.0
soupsieve==1.6
swagger-spec-validator==2.4.1
tzlocal==1.5.1
urllib3==1.24.1
Werkzeug==0.14.1
